Bind mongodb

+++ services/tools/models.go
@@ -1,14 +1,14 @@
 package tools
 
 type SearchRequest struct {
-	Start  int64             `json:"start" validate:"-"`
-	Length int64             `json:"length" validate:"-"`
+	Start  int             `json:"start" validate:"-"`
+	Length int             `json:"length" validate:"-"`
 	Query  map[string]string `json:"query" validate:"-"`
 }
 
 type Paginate struct {
-	Length int64       `json:"length"`
-	Offset int64       `json:"offset"`
-	Total  int64       `json:"total"`
+	Length int       `json:"length"`
+	Offset int       `json:"offset"`
+	Total  int       `json:"total"`
 	Items  interface{} `json:"items"`
 }
